Tag archives
- adds a plugin file (or two) that handle generation of tag archive pages

In the plugin directory there is now a file that will generate an archive page for each tag found in
the front matter for any post. The fun part (or only bug I can think of) is that the generator script
is case sensitive and will ignore any tags that are of different capitalization from the first one
found. Thus, this commit also includes changes to some of the posts imported from tumblr to make the
imported tag names consistent across the site and ensure they are included in the archive.