Rips out jekyll-get
The `jekyll-get` plugin was slowing down builds, duplicating (or
triplicating) data, and recently messing up alphabetization. These data
aren't changing frequently enough to require a refresh on every site
build and the complications are outweighing the benefits at this point.
There might be a future need for constant refreshing and in that case we
might consider a cron job that downloads fresh copies periodically.
Until then, we need to make sure that future team photo additions are
accompanied by a refresh of these data.