It's a little random, shaky even, but not broken!
If we read out of the "team" collection files to generate the index, we
get this somewhat alphabetical, somewhat random ordering on the
homepage. We can easily change this to read out of the data file
instead. But it works! Each person should at least have their name,
photo on the index, and full name and photo on their profile