It was human after all
Turns out `display: inline-block` was the reason for the text alignment
issues on the `h1` elements in the bio divs. Figured this out after
noticing that on smaller screens the bios floated nicely alongside one
another without any problems. The only difference between smaller and
larger screens was, on smaller screens, `display` was `block` and we
were floating left and clearing, thus preventing the blocks from
overlapping one another.