moved up call to action, removed bios
So call to action doesn't get lost in the page. Removed bios so we are not duplicating content on the team page and also so /consulting is consistent with other internal pages. In the next iteration we will add a link to the team page sorted just for consulting. There may be other design ways to get the team on /consulting as well.