Lower-cases older tags
Tags should always be lower case, this is because the Jekyll archives plugin
lower cases them automatically when building the pages but still sees
them as different tags. If there both tags, "18f" and "18F" exist, the
archive page for `tag/18f` will get generated with only the posts for
one of them included. It's not clear whether it's the first one or the
second.