Tests will fail until morale improves
These tests fail! And that's good! Because we have missing frontmatter
all over the place and we need to fix it so we don't have funny looking
posts in production. I propose we leave this open until we've fixed all
the frontmatter problems and then merge away! (Also I should comment
this code a bit.)