Applies the new standard template to /consulting.
In theory this template is ready for all the pages but we'll want to proof them to look for design inconsistencies before going primetime.

Changes from the /press and /team page commits: you can now specify a "subtitle" on a page and it will show up in the second line of the "headline" for any document using the layout "page."

Outstanding changes remaining: create a style for posts to use the headline. This should be relatively straightforward at this point.