Permit code samples in blog posts that will run against a Jupyter server
Using O'Reilly's `Thebe` project, allow code samples with a `data-executable`
attribute to run live against a Project Jupyter server
(running with https://github.com/jupyter/tmpnb).

An example of how to use is in the (purely for demo purposes)
scratch post at _posts/2015-10-07-how-we-blog-with-jupyter.md.

The Jupyter server is not yet spun up.  It is expected at jupyter.18f.gov:8000.
Until it is up, these scripts won't actually work.