Copy change for mobile optimization
"Sign up to subscribe to our newsletter" is too long and spills over the button on mobile (320px). I think "Subscribe to our newsletter" adequately conveys the correct information to people, and also makes the button work for mobile. The alternative that I tried was reducing the font-size to 0.9em, but this is a better solution IMO.