Correct factual error
"87% of large government IT projects fail" Isn't true. We can say, "87% of large government IT projects don't succeed," "29% of large government IT projects fail," or "Only 13% of large government IT projects succeed". [The underlying data is:  13% successful, 58% challenged and 29% failed] This is also from 2015 so I feel like it is getting stale.

Also removed whitespace