Add site-wide pa11y checking
This commit adds site-wide pa11y checking. If a "global" folder (like _includes or _sass) sees a file change, we add a few files from several folders across the site, and include them in the list of files for pa11y to check.

This also switches file "differs" based on the environment. If CI=true, then we diff the files from the last commit. Otherwise, we assume we're developing locally and get the files currently changed.

This feels like draft code, the design feels off. There's constants set at file load/runtime, and with the new class involved, the file is a lot harder to read.