Remove conditions on the Jekyll build
The pa11y_targets file is always going to be created by the jekyll
build and it doesn't seem harmful to go ahead and store the list of
changed files into the env variable regardless of whether this is a
full site scan being done nightly or if it's in response to a commit
from GitHub on a pull request. This simplifies the configuration
logic.