Start replatforming branch
In order to maintain maximum parity between 18f.gsa.gov and 18F/guides, we are copying most of the 11ty setup from 18F/guides to serve as a starting point.

This commit does not result in a successful build.

Plus, many files — like assets, pa11y config, and others — still need to be reconciled.

This commit:
- Removes many Ruby- and Jekyll-specific files
- Migrates many files from 18F/guides, moving them into roughly the right folders
- Adds a placeholder README based off 18F/guides