Finish migrating authors
This commit:
- Removes "author published" status checks
- Re-implements team_link and team_photo filters
- Re-implements an "authors' matching posts" filter.
- Refactors the author post display
  - Refactors layout into partials
  - Refactors various author / post display logic
  - Removes unnecessary (hopefully) context mutation
- Fixes author data
  - Adds missing team member data
  - Renames authors.csv -> team_members.csv because global data was polluting post.data.authors

The consequence of removing the `published` is that we'll have 63 pages for unpublished authors, most/all of whom are team members featured in the "we asked 100 colleagues" posts.

These pages are unlinked so no one is likely to stumble upon them by accident. Perhaps we can add a robots.txt entry for them later so they're not indexed.

This doesn't have a meaningful impact on performance.