Review layouts
This commit:

- Removes unused layouts and partials
- Reviews all layouts and partials for correct data
  - Removes `include.` references, not part of 11ty
- Removes defunct .about.yml data
- Minor tweaks to code spacing, layout, indentation

While working on this, I noticed some duplicate includes, like header is basically the same as usa-banner. I think that's why we don't have a menu: I suspect the `header` file in Jekyll was the top menu, but got overwritten by the "this is a .gov site" banner, which is called usa-banner in Jekyll but header in 18F/guides.