Remove admin files
The Guides and 18f.gsa.gov intended to make it possible to edit and publish using a CMS like Netlify, which these files would have configured.

However, we don't presently use a CMS, nor is there a clear immediate pathway to get one, so we'll remove these files for now.