Fix testimonials and reorganize pages
The only substantial issue found in diffing was that testimonials had gone missing — we hadn't yet converted those quotes.

This also showed some pages were missing layouts.

While doing this, it became clear that some pages belong together, like all the offerings linked from work-with-us, as well as the categories in the footer.