Update glob to 9.x.x and 11ty SVG sprite plugin to 2.x.x
Snyk tried to update our glob dependency to 9.x but the build was failing. It took quite a bit of digging to find what I think was the culprit which was v 1.3.1 of the Eleventy SVG sprite plugin that we use to render USWDS icons on the 18f site.

eleventy-plugin-svg-sprite v1.3.1 does not list any version of `glob` as a dependency in its own `package.json` file but it makes use of `glob` in its `src/SVGSprite` file. The code goes on to call `util.promisify(glob)` which lines up with the error I was getting before updating the plugin version.